pyproject.toml
[project]
name = "poc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
    "ultralytics>=8.3.203",
    "pillow>=10.4.0",
    "streamlit>=1.50.0",
]
